feat: SSG にした
